started to improve folder structure - made some small adjustments to style - making code moe DRY
