+++ client/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Birmingham Chiropractic",
+  "name": "Birmingham Chiropractic - 264 Alcester Rd South, King's Heath, B14 6DR",
   "icons": [
     {
       "src": "favicon.ico",
