--- client/src/App.test.js
@@ -1,3 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import App from './App';
-
