+++ client/src/custom.scss
@@ -1,4 +1,4 @@
  $dark: #12485a;
- $light:#f6f9fa;
+ $light: #f9f7f7;
 
 @import '~bootstrap/scss/bootstrap.scss';
