+++ client/src/pages/about/about.css
@@ -19,7 +19,7 @@
 }
 
 .about-header {
-  font-size: 4.5em;
+  font-size: 4rem;
 }
 
 .josefinSans-text {
