+++ client/src/pages/home/WhatWeDoNav.js
@@ -1,6 +1,9 @@
-import foot from "../images/foot.png";
-import spine from "../images/spine-model-.jpg";
-import runner from "../images/woman-running.jpg";
+import foot from "../../images/foot.png";
+import spine from "../../images/spine-model-.jpg";
+import runner from "../../images/woman-running.jpg";
+import team from "../../images/team.jpg";
+import contact from "../../images/phone.jpg";
+import testimonials from "../../images/new-imgs/thumbUp.png";
 import LinkCard from "./LinkCard";
 
 function WhatWeDoNav() {
@@ -23,10 +26,30 @@ function WhatWeDoNav() {
       text: "Sports Remedial",
       key: "sports",
     },
+    {
+      image: team,
+      altText: "Doctor",
+      text: "Meet the Team",
+      key: "meetTeam",
+      link: "about",
+    },
+    {
+      image: testimonials,
+      altText: "Testimonials",
+      text: "Testimonials",
+      key: "Testimonials",
+      link: "testimonials",
+    },
+    {
+      image: contact,
+      altText: "Contact Us",
+      text: "Contact Us",
+      key: "contact",
+    },
   ];
 
   return (
-    <div className="container-md pb-5">
+    <div className="container-md my-5 pb-5">
       <div className="row gy-3">
         {cards.map((it) => (
           <div className="col-12 col-md-4" key={it.key}>
@@ -34,6 +57,7 @@ function WhatWeDoNav() {
               image={it.image}
               text={it.text}
               altText={it.altText}
+              link={it.link}
             ></LinkCard>
           </div>
         ))}
